% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windninjr-package.R
\docType{package}
\name{windninjr-package}
\alias{windninjr}
\alias{windninjr-package}
\title{windninjr: Helper Functions for Running WindNinja}
\description{
Helper files for running WindNinja.
}
\author{
\strong{Maintainer}: Ben Raymond \email{ben.raymond@aad.gov.au}

}
\keyword{internal}
