% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_files.R
\name{wn_config_domain_average}
\alias{wn_config_domain_average}
\title{Generate a configuration file with domain-average wind initialization}
\usage{
wn_config_domain_average(
  input_speed,
  input_speed_units = "mps",
  input_direction,
  input_wind_height = 10,
  units_input_wind_height = "m",
  momentum_flag = TRUE,
  number_of_iterations = 300L,
  output_wind_height = 10,
  units_output_wind_height = "m",
  output_speed_units = "mps",
  ...
)
}
\arguments{
\item{input_speed}{numeric: vector of wind speeds}

\item{input_speed_units}{string: units of input wind speed ("mps", "mph", "kph", "kts")}

\item{input_direction}{numeric: vector of input directions in degrees (direction that wind is blowing towards - e.g. 135 means that the wind is blowing towards the northwest)}

\item{input_wind_height}{numeric: height of input wind speed above the vegetation}

\item{units_input_wind_height}{string: units of input wind height ("ft", "m")}

\item{momentum_flag}{logical: use momentum solver?}

\item{number_of_iterations}{numeric: number of iterations for momentum solver}

\item{output_wind_height}{numeric: height of input wind speed above the vegetation}

\item{units_output_wind_height}{string: units of output wind height ("ft", "m")}

\item{output_speed_units}{string: units of input wind speed ("mps", "mph", "kph", "kts")}

\item{...}{: additional parameters, see \code{\link{wn_config_common}}}
}
\value{
The path to the configuration file
}
\description{
Generate a configuration file with domain-average wind initialization
}
\examples{
\dontrun{
  my_config <- wn_config_domain_average(elevation = wn_demo_file("missoula_valley_elevation"),
                                        input_speed = 10, input_direction = 270)
}

}
\seealso{
\code{\link{wn_config_common}}
}
