% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{wn_demo_file}
\alias{wn_demo_file}
\title{Demonstration files bundled with the package}
\usage{
wn_demo_file(what = "missoula_valley_elevation")
}
\arguments{
\item{what}{string: which file?
\itemize{
\item "missoula_valley_elevation" - the Missoula Valley elevation geotiff file bundled with WindNinja itself
}}
}
\value{
The path to the file
}
\description{
Demonstration files bundled with the package
}
