% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{wn_read}
\alias{wn_read}
\title{Convenience function for reading WindNinja output files}
\usage{
wn_read(path, pattern, vars = c("vel", "ang"), format = "asc")
}
\arguments{
\item{path}{string: path to output directory}

\item{pattern}{string: an optional regular expression (e.g. \code{"_270_10_100m_"}) passed on to grep() to filter files}

\item{vars}{character: a character list of variables to return}

\item{format}{string: output format to look for}
}
\value{
A raster stack
}
\description{
Convenience function for reading WindNinja output files
}
