% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{wn_run}
\alias{wn_run}
\title{Run WindNinja}
\usage{
wn_run(config, output_dir = tempfile(), create_output_dir = TRUE)
}
\arguments{
\item{config}{list: configuration object as returned by e.g. \code{\link{wn_config_domain_average}}}

\item{output_dir}{string: path to output directory}

\item{create_output_dir}{logical: if \code{TRUE} and the output directory does not exist, create it}
}
\value{
A list
}
\description{
Run WindNinja
}
\examples{
\dontrun{
  demfile <- wn_demo_file("missoula_valley_elevation")
  my_config <- wn_config_domain_average(elevation = demfile,
                                        input_speed = 10, input_direction = 270)
  res <- wn_run(my_config)
}

}
